MCTS with parallelism
MCTS with parallelism. Result: not very good... Too much memory usage and even if the game were to be optimized with bit board and other techniques, it wouldn't be able to simulate 1.2 million games per hour, which is apparently the benchmark for alpha-zero. Result is disappointing, but the code works fine and theoretically it's the same Monte Carlo implementation one would find anywhere else. This was one of the more exciting personal projects I've done in a while albeit it doesn't include neural network like alpha-zero.